# api/openapi.yaml
info:
  title: SakuinSeizouki API
  version: 0.1.0
tags:
- name: Pdf
- name: Word List
- name: Index
servers:
- url: http://localhost:34200/
paths:
  /index/get:
    post:
      parameters: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetIndexOut'
          description: Successful response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
          description: Validation error
      tags:
      - Index
      summary: Get Index
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIndexOut'
  /word-list/:
    get:
      parameters:
      - in: query
        name: word_list_path
        schema:
          type: string
        required: false
      - in: query
        name: sheet_name
        schema:
          type: string
        required: false
      - in: query
        name: start_cell
        schema:
          type: string
        required: false
      - in: query
        name: end_cell
        schema:
          type: string
        required: false
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetWordListOut'
          description: Successful response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
          description: Validation error
      tags:
      - Word List
      summary: Get Word List
  /index/create:
    post:
      parameters: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateIndexOut'
          description: Successful response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
          description: Validation error
      tags:
      - Index
      summary: Create Index
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIndexIn'
  /pdf/page-types:
    get:
      parameters:
      - in: query
        name: pdf_path
        schema:
          type: string
        required: false
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetPageTypesOut'
          description: Successful response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
          description: Validation error
      tags:
      - Pdf
      summary: Open Pdf
  /pdf/page-image:
    get:
      parameters:
      - in: query
        name: pdf_path
        schema:
          type: string
        required: false
      - in: query
        name: page_number
        schema:
          type: integer
        required: false
      responses:
        '200':
          content:
            image/png:
              schema: {}
          description: Successful response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
          description: Validation error
      tags:
      - Pdf
      summary: Get Page Image
  /temp/{path}:
    get:
      parameters:
      - in: path
        name: path
        schema:
          type: string
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful response
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Not found
      summary: Get File
openapi: 3.0.3
components:
  schemas:
    ValidationError:
      properties:
        detail:
          type: object
          properties:
            <location>:
              type: object
              properties:
                <field_name>:
                  type: array
                  items:
                    type: string
        message:
          type: string
      type: object
    HTTPError:
      properties:
        detail:
          type: object
        message:
          type: string
      type: object
    WordPages:
      type: object
      properties:
        word:
          type: string
        pages:
          type: array
          items:
            type: integer
    CreateIndexOut:
      type: object
      properties:
        word_pages:
          type: array
          items:
            $ref: '#/components/schemas/WordPages'
        missing_pages:
          type: array
          items:
            type: integer
        missing_words:
          type: array
          items:
            type: string
    GetIndexOut:
      type: object
      properties:
        url:
          type: string
    GetWordListOut:
      type: object
      properties:
        word_list:
          type: array
          items:
            type: string
    Annotation:
      type: object
      properties:
        x:
          type: integer
        y:
          type: integer
        width:
          type: integer
        height:
          type: integer
        group_index:
          type: integer
    PageTypeDetail:
      type: object
      properties:
        annotations:
          type: array
          items:
            $ref: '#/components/schemas/Annotation'
        page_numbers:
          type: array
          items:
            type: integer
    CreateIndexIn:
      type: object
      properties:
        list_path:
          type: string
        pdf_path:
          type: string
        sheet_name:
          type: string
        start_cell:
          type: string
        end_cell:
          type: string
        page_types:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/PageTypeDetail'
    PageType:
      type: object
      properties:
        width:
          type: integer
        height:
          type: integer
        page_numbers:
          type: array
          items:
            type: integer
        type:
          type: integer
    GetPageTypesOut:
      type: object
      properties:
        page_types:
          type: array
          items:
            $ref: '#/components/schemas/PageType'